Backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.29
passlib[bcrypt]==1.7.4
python-jose==3.3.0
bcrypt==4.1.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9